% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_countries}
\alias{all_countries}
\title{View All Countries}
\usage{
all_countries()
}
\value{
A vector
}
\description{
This function returns names of all the countries in the world.
}
\details{
This function allows the user to create a vector of all countries.
}
\author{
Michael Mbajwa
}
