% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_indicators}
\alias{all_indicators}
\title{All World Bank Indicators}
\usage{
all_indicators()
}
\value{
A vector
}
\description{
This function extracts vector of World Bank indicators.
}
\details{
This function allows the user to create a vector of World Bank indicators.
}
\author{
Michael Mbajwa
}
