% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_indicators_like}
\alias{all_indicators_like}
\title{All Indicators like String}
\usage{
all_indicators_like(like)
}
\arguments{
\item{like}{A string the indicator name should contain.}
}
\value{
A vector
}
\description{
This function returns all indicators that contain string provided.
}
\details{
This function returns all indicators that contain strings provided
}
\examples{
# See all indicators that contain health in them
all_indicators_like("health")
}
\author{
Michael Mbajwa
}
