% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{country_indicator}
\alias{country_indicator}
\title{This function returns World Bank Indicators for a specified country.}
\usage{
country_indicator(country_name, country_code, indicator_contains, year)
}
\arguments{
\item{country_name}{The name of the country whose details the user wants to extract.}

\item{country_code}{The country code of the country.}

\item{indicator_contains}{A string that world bank indicator should contain.}

\item{year}{The year you are interested in. Must be a string.}
}
\value{
A data frame
}
\description{
This function returns World Bank Indicators for a specified country.
}
\details{
This function allows the user to create a data frame containing key world bank indicators containing a
certain string. Indicators are filtered down to the year and country name.
Either one country code or country name must be provided.
See the example below for its usage.
}
\examples{
# See country indicator containing health for America using country name
country_indicator(country_code = "usa", indicator_contains = "health", year = "2002")
}
\author{
Michael Mbajwa
}
