% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{country_info}
\alias{country_info}
\title{Helper function for country_key_details.}
\usage{
country_info(col_name, col_value)
}
\arguments{
\item{col_name}{The column name to be extracted from the loaded data frame.}

\item{col_value}{The value which is used to filter col_name}
}
\value{
A data frame
}
\description{
Helper function for country_key_details.
}
\details{
This is a reusable function that allows other functions to create a filtered data frame with
a specific column of interest having specified values.
}
