% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{country_key_details}
\alias{country_key_details}
\title{Return Key Country Info}
\usage{
country_key_details(country_name, country_code)
}
\arguments{
\item{country_name}{The name of the country whose details the user wants to extract.}

\item{country_code}{The country code of the country.}
}
\value{
A data frame
}
\description{
This function returns important information regarding specified country.
}
\details{
This function allows the user to create a data frame containing general overview of a country.
The information returned contains Country Name, Country Code, Currency, Unit, Income Group,
Head of State, and gender of head of state.
Either one of the parameters must be provided.
}
\examples{
# See country details for America using country code
library("CountryIndicatorsPkg")
country_key_details(country_code = "usa")
# See country details for Nigeria using country Name
country_key_details(country_name = "nigeria")
}
\author{
Michael Mbajwa
}
