% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{country_leader}
\alias{country_leader}
\title{This function returns the country leader for the specified country.}
\usage{
country_leader(country_name, country_code)
}
\arguments{
\item{country_name}{The name of the country whose country_leader the user wants to extract.}

\item{country_code}{The country code of the country.}
}
\value{
A data frame
}
\description{
This function returns the country leader for the specified country.
}
\details{
This function allows the user to create a data frame containing general information of a country's leader.
The information returned contains, Head of State name, and gender of head of state.
Either one of the parameters must be provided.
See the examples below for its usage.
}
\examples{
# See country leader details for America using country code
country_leader(country_code = "usa")
# See country leader details for Luxembourg using country name
country_leader(country_name = "luxembourg")
}
\author{
Michael Mbajwa
}
