% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalCompleteData.R
\docType{data}
\name{FinalCompleteData}
\alias{FinalCompleteData}
\title{Data of global world bank development indicators in wide format}
\format{
A tibble with 151305 rows and 75 columns:
\describe{
\item{Country_Name}{chr Country Name}
\item{Country_Code}{chr Code of Country}
\item{Indicator_Name}{chr World Bank Development Indicator}
\item{Indicator_Code}{chr World Bank Development Indicator code for indentification}
\item{19..}{chr Year}
\item{2..}{chr Year}
\item{Short_Name}{chr Short name for country}
\item{Table_Name}{chr Table name for country}
\item{Long_Name}{chr long name for country}
\item{Currency_Unit}{chr currency unit used in country}
\item{Region}{chr Region of the world country is located in}
\item{Income_Group}{chr income group country belongs in}
\item{Head_of_state}{chr name of the head of state}
\item{Head_of_government}{chr name of the head of government}
\item{Gender_HoS}{chr gender of head of state}
}
}
\source{
\url{http://databank.worldbank.org/data/download/WDI_excel.zip}
}
\usage{
FinalCompleteData
}
\description{
To be used in this package
}
\keyword{datasets}
