% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{indicators_info}
\alias{indicators_info}
\title{Helper Function}
\usage{
indicators_info(col_name, col_value, indicator_contains, year)
}
\arguments{
\item{col_name}{The column name to be extracted from the loaded data frame.}

\item{col_value}{The value which is used to filter col_name.}

\item{indicator_contains}{A string that indicator should contain.}

\item{year}{The year you are interested in. Must be a string.}
}
\value{
A data frame
}
\description{
Helper function for country_indicator
}
\details{
This is a reusable function to be used by other functions.
}
