% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_plot_countries}
\alias{make_plot_countries}
\title{Plot all countries}
\usage{
make_plot_countries(map_title = "World Map")
}
\arguments{
\item{map_title}{A string containing desired title for the plot}
}
\value{
A plot
}
\description{
This function returns a plot of all the countries we have indicators for.
}
\details{
This function returns a plot of all the countries we have indicators for.
}
\examples{
make_plot_countries()
}
\author{
Michael Mbajwa
}
