% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_country}
\alias{map_country}
\title{Country map}
\usage{
map_country(all_country_details)
}
\arguments{
\item{all_country_details}{A dataframe returned from the function country_key_details("nigeira").}
}
\value{
A ggplot
}
\description{
This function returns a plot of specified country with key details provided.
}
\details{
This function returns a plot of specified country with key details provided.
}
\examples{
# See all indicators that contain health in them
map_country(all_country_details=country_key_details("nigeria"))
}
\author{
Michael Mbajwa
}
