% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyAppData.R
\docType{data}
\name{ShinyAppData}
\alias{ShinyAppData}
\title{Data of global world bank development indicators in long format}
\format{
A tibble with 9,380,900 rows and 4 columns:
\describe{
\item{Country_Name}{chr Country Name}
\item{Indicator_Name}{chr World Bank Development Indicator}
\item{Year}{chr Year}
\item{values}{chr indicator value}
}
}
\usage{
ShinyAppData
}
\description{
To be used in the shiny app developed in the reproducible pipeline: https://github.com/Michael-Mbajwa/CountryIndicatorsPipeline
}
\keyword{datasets}
